requirements.txt
# ESP32 flashing and communication tools
esptool>=4.0.0

# Serial communication for device detection and monitoring
pyserial>=3.5

# HTTP requests for API communication with DinoCore
requests>=2.25.0

# Bluetooth LE support for QC testing (optional)
bleak>=0.19.0

# Firebase/Firestore integration for data logging (optional)
firebase-admin>=6.0.0

# Selenium for web automation (for MeloTTS integration)
selenium>=4.0.0

# Flask for API bridge (n8n integration)
flask>=2.0.0

# Note: The following are built-in Python modules and don't need to be installed:
# - subprocess, sys, time, os, json, argparse, glob, platform, readline, atexit
# - tkinter (GUI framework, included with Python)
